refactor: move user domain ports into ferriskey-domain
