+++ core/Cargo.toml
@@ -8,7 +8,7 @@ edition.workspace = true
 [dependencies]
 ferriskey-abyss = { path = "../libs/ferriskey-abyss", features = ["mock"] }
 ferriskey-organization = { path = "../libs/ferriskey-organization", features = ["mock"] }
-ferriskey-domain = { path = "../libs/ferriskey-domain" }
+ferriskey-domain = { path = "../libs/ferriskey-domain", features = ["mock"] }
 ferriskey-security = { path = "../libs/ferriskey-security", features = ["mock"] }
 ferriskey-trident = { path = "../libs/ferriskey-trident" }
 ferriskey-aegis = { path = "../libs/ferriskey-aegis" }
