+++ libs/ferriskey-domain/Cargo.toml
@@ -12,6 +12,11 @@ utoipa = { version = "5.4.0", features = ["chrono", "uuid"] }
 uuid = { version = "1.20.0", features = ["serde", "v4", "v7"] }
 thiserror = "2.0.12"
 tracing = "0.1"
+mockall = { version = "0.14.0", optional = true }
 
 [dev-dependencies]
 serde_json = "1.0.149"
+mockall = "0.14.0"
+
+[features]
+mock = ["mockall"]
