+++ libs/ferriskey-domain/src/user/mod.rs
@@ -1,3 +1,4 @@
 pub mod commands;
 pub mod entities;
+pub mod ports;
 pub mod value_objects;
